rename GridBuilder to GridDescriptor
